requirements.txt
APScheduler==3.11.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
bcrypt==4.3.0
blinker==1.8.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.4
cssselect2==0.7.0
defusedxml==0.7.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
html5lib==1.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
lxml==6.0.0
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==1.24.4
oscrypto==1.3.0
pandas==2.0.3
pillow==10.4.0
pycparser==2.22
pyHanko==0.27.1
pyhanko-certvalidator==0.26.8
PyMySQL==1.1.1
pypdf==5.8.0
pypng==0.20220715.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.4.2
requests==2.32.4
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
svglib==1.5.1
threadpoolctl==3.5.0
tinycss2==1.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.0.6
xhtml2pdf==0.2.17
zipp==3.20.2
